OAuth/requirements.txt
anyio==4.2.0
asgiref==3.7.2
azure-common==1.1.28
azure-communication-email==1.0.0
azure-core==1.29.6
azure-mgmt-core==1.4.0
azure-storage-blob==12.19.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
defusedxml==0.8.0rc2
Django==4.2.3
django-cors-headers==4.3.1
django-filter==23.5
django-picklefield==3.1
django-q2==1.6.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
idna==3.6
isodate==0.6.1
msrest==0.7.1
numpy==1.26.2
oauthlib==3.2.2
pandas==2.1.4
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
xlrd==2.0.1
